Backened/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
google-generativeai==0.3.0
transformers==4.35.0
torch==2.1.0
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
